Test Console EventHandlers.
Tests for Console input and output event handlers have been written and
the classes have been updated to pass. Output prints to the System.out and
is done. Console input is split into macro and verbose modes. macro mode
looks for a single input and returns when it is obtained. The other class
creates a thread and continually checks for new input. When it there is
new input, it is delivered to the Reactor as an event. Can this be updated
to deliver the event to the reactor on the main thread? Perhapse the
Reactor is supposed to have its own main loop that handles the input after
it is received. This may also symchronize the events dispatched from the
reactor.