Evaluate returns ExpressionTree. New EvaluatedState object.
Evaluate returns an expression tree instead of a Double. The expression
Tree is created with a context. Since the context, when state allows,
evaluates uses itself as the evalate context param. The context should
record this as its current evaluation tree, but for this to work correctly
there needs to be an EvaluatedState.