Update EventHandlers, Macro and Help command, plus tests.
Updated the macro and verbose command event handlers. Created a new
interface HelpEventHandler and a help command. The help command raises a
help event with the reactor.
The context now has a new boolean. MacroCommands checks the boolean to see
if the last command threw an exception. if so the macro stops execution.
New EventHandlerException subclass for when an Reactor.handleEvent() fails
to find a handler.